week1_project/displaytalent/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
